// app/src/generated/BypassLink.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BypassLink = { "type": "wormhole", a: number, b: number, } | { "type": "gateway", system: number, active: boolean, } | { "type": "l_gate", system: number, } | { "type": "other", system: number, kind: string, };